pyproject.toml
[project]
name = "carga-termica-bovinos"
version = "0.1.0"
description = "Empirical bovine thermal comfort analysis in psychrometric space"
authors = [
    { name = "João Gerd Zell de Mattos" },
    { name = "Cleicione Moura de Oliveira" },
    { name = "Rafael Augusto Satrapa" },
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "scipy",
    "pydantic",
    "pyyaml",
    "alphashape",
    "shapely",
    "seaborn",
    "pyarrow",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
